Data_Display has been coded which allows for the displayu to be printed on the console and has a file that opoens and constantly update the log

+++ coen320-ATCsim/Aircraft.cpp
@@ -104,7 +104,7 @@ void Aircraft::setInAirspaceFlag(bool flag) {
 	this->inAirspaceFlag = flag;
 }
 
-int Aircraft::getAnticipatedx()
+double Aircraft::getAnticipatedx()
 {
 	return this->anticipatedx;
 }
